+++ src/Layout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {Navigation} from "./Navigation";
+import {Navigation} from "./components/Navigation";
 
 const Layout = ({children} : {children:React.ReactElement}): React.ReactElement => {
     return(
